About halfway there, but the heavy lifting is done
