--- test/ly/stealth/xmlavro/sax/outputs/expected_output_one
@@ -1,22 +1,0 @@
-startElement
-startElement
-startElement
-endElement
-startElement
-endElement
-startElement
-endElement
-startElement
-endElement
-endElement
-startElement
-startElement
-endElement
-startElement
-endElement
-startElement
-endElement
-startElement
-endElement
-endElement
-endElement
